cleanup annotation view code
